Deleted the ShortestPathScript and put its functionality in NipperTour. Attached the NipperTour.cs on the PathCreator game object in Unity and removed this script from the F2 game object. Now the PathCreator draws the line for the shortest path in the Update() method.

+++ GitHub.CSharpTests4/Assets/Scripts/Map/NodeTests.cs
@@ -22,8 +22,8 @@ namespace Assets.Scripts
                 g.InsertNewVertex(n1);
                 g.InsertNewVertex(n2);
 
-                n1.addAdjacentNode(new Dictionary<Node, double>() {{n2, 4}});
-                n2.addAdjacentNode(new Dictionary<Node, double>() {{n1, 3}});
+                n1.addAdjacentNode(new Dictionary<Node, float>() {{n2, 4}});
+                n2.addAdjacentNode(new Dictionary<Node, float>() {{n1, 3}});
                 Assert.True(n1.hasAdjacentNode(n2),
                     "This tests if hasAdjacentNode returns true if a node is adjacent to a given one.");
             }
